fix: refresh rate fixes
