+++ src/client/src/app/(playground)/dashboard/llm/requests-per-time.tsx
@@ -17,7 +17,6 @@ import { COLORS } from "../../../../../colors";
 import IntermediateState from "@/components/(playground)/intermediate-state";
 import { toast } from "sonner";
 import { getFilterParamsForDashboard } from "@/helpers/filter";
-import { useRefreshRate } from "@/utils/hooks/useRefreshRate";
 
 export default function RequestsPerTime() {
 	const filter = useRootStore(getFilterDetails);
@@ -36,8 +35,6 @@ export default function RequestsPerTime() {
 			},
 		});
 	}, [filter]);
-
-	useRefreshRate(fetchData);
 
 	useEffect(() => {
 		if (
