+++ src/client/src/app/(playground)/dashboard/vector/operations.tsx
@@ -15,7 +15,6 @@ import {
 	XAxis,
 	YAxis,
 } from "recharts";
-import { useRefreshRate } from "@/utils/hooks/useRefreshRate";
 
 export default function Operations() {
 	const filter = useRootStore(getFilterDetails);
@@ -30,8 +29,6 @@ export default function Operations() {
 			responseDataKey: "data",
 		});
 	}, [filter]);
-
-	useRefreshRate(fetchData);
 
 	useEffect(() => {
 		if (
