+++ src/client/src/components/(playground)/request/request-filter.tsx
@@ -19,7 +19,6 @@ import useFetchWrapper from "@/utils/hooks/useFetchWrapper";
 import { useCallback, useEffect, useState } from "react";
 import { toast } from "sonner";
 import { FilterConfig, FilterType } from "@/store/filter";
-import { useRefreshRate } from "@/utils/hooks/useRefreshRate";
 
 const DynamicFilters = ({
 	isVisibleFilters,
@@ -90,8 +89,6 @@ const DynamicFilters = ({
 			},
 		});
 	}, [filter]);
-
-	useRefreshRate(fetchConfig);
 
 	useEffect(() => {
 		if (
