+++ src/client/src/components/(playground)/stat-card.tsx
@@ -9,7 +9,6 @@ import { Skeleton } from "@/components/ui/skeleton";
 import { LucideIcon, TrendingDown, TrendingUp } from "lucide-react";
 import IntermediateState from "./intermediate-state";
 import { getFilterParamsForDashboard } from "@/helpers/filter";
-import { useRefreshRate } from "@/utils/hooks/useRefreshRate";
 
 type StatCardProps = {
 	heading?: string;
@@ -57,8 +56,6 @@ const StatCard = memo(
 				responseDataKey: "data[0]",
 			});
 		}, [filter, url]);
-
-		useRefreshRate(fetchData);
 
 		useEffect(() => {
 			if (
