+++ src/client/src/store/filter.ts
@@ -1,4 +1,4 @@
-import { set } from "lodash";
+import { merge, set } from "lodash";
 import { addDays, addMonths, addWeeks } from "date-fns";
 import { lens } from "@dhmk/zustand-lens";
 
@@ -61,11 +61,11 @@ export interface FilterConfig {
 	traceTypes: string[];
 }
 
-function getTimeLimitObject(
+export function getTimeLimitObject(
 	value: string,
 	keyPrefix: string,
 	extraParams?: any
-) {
+): unknown {
 	let object = {};
 	if (value === TIME_RANGE_TYPE["24H"]) {
 		const currentDate = new Date();
@@ -97,7 +97,10 @@ function getTimeLimitObject(
 const INITIAL_FILTER_DETAILS: FilterType = {
 	timeLimit: {
 		type: DEFAULT_TIME_RANGE,
-		...getTimeLimitObject(DEFAULT_TIME_RANGE, ""),
+		...(getTimeLimitObject(DEFAULT_TIME_RANGE, "") as {
+			end: Date;
+			start: Date;
+		}),
 	},
 	limit: DEFAULT_LIMIT,
 	offset: 0,
@@ -120,7 +123,10 @@ export const filterStoreSlice: FilterStore = lens((setStore, getStore) => ({
 		let resetConfig = false;
 		switch (key) {
 			case "timeLimit.type":
-				object = getTimeLimitObject(value, "timeLimit.", extraParams);
+				object = getTimeLimitObject(value, "timeLimit.", extraParams) as {
+					end: Date;
+					start: Date;
+				};
 				resetConfig = true;
 				break;
 			case "limit":
@@ -146,9 +152,8 @@ export const filterStoreSlice: FilterStore = lens((setStore, getStore) => ({
 
 		setStore({
 			details: {
-				...getStore().details,
+				...merge(getStore().details, object),
 				selectedConfig: resetConfig ? {} : getStore().details.selectedConfig,
-				...object,
 			},
 			config: resetConfig ? undefined : getStore().config,
 		});
